removed nisa 2 and tesco 3 parsers + unit tes coverage
