+++ app/services/parser-model.js
@@ -7,11 +7,9 @@ const ParserModel = {
   BANDM1: "bandm-1",
   TESCO1: "tesco-1",
   TESCO2: "tesco-2",
-  TESCO3: "tesco-3",
   FOWLERWELCH1: "fowlerwelch-1",
   NISA1: "nisa-1",
   NISA2: "nisa-2",
-  NISA3: "nisa-3",
   BUFFALOAD1: "buffaload-1",
   CDS1: "cds-1",
   COOP1: "coop-1",
